# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passwordless Multi-Auth API",
    "description": "Passwordless Multi-Auth Platform API - Passwordless Authentication\n\nThis API provides comprehensive endpoints for:\n- MagicLink-based passwordless authentication\n- User management and security operations\n- Role and authority management\n- Push notifications\n- TOTP/MFA verification\n\n## Authentication Flow\n1. Call `POST /api/auth/email-magiclink/send` with user email\n2. User opens the emailed link in a browser: `GET /api/auth/verify?token={token}`\n3. The server redirects (302) to the app scheme `passwordless://auth/callback?code={code}` (or `?status=totp_required` when MFA is enabled, `?status=error&reason=...` on failure)\n4. Trade the one-time code for a token pair via `POST /api/auth/exchange` (if TOTP was required, call `POST /api/auth/totp/verify` instead)\n5. Use the returned JWT in the Authorization header: `Bearer {token}`\n6. Renew with `POST /api/auth/refresh` (single-use rotation); revoke with `POST /api/auth/logout`\n\n## Flutter Integration\nFor Flutter apps, use the `/api/auth/email-magiclink/send` endpoint\nfor mobile authentication. The MagicLink can be handled via\ndeep linking in your Flutter app.\n",
    "contact": {
      "name": "Passwordless Multi-Auth Team",
      "email": "api@example.com"
    },
    "license": {
      "name": "Private",
      "url": "https://github.com/ersincivi/spring-boot-passwordless-multi-auth"
    },
    "version": "1.0.0"
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/email-otp/verify": {
      "post": {
        "tags": [
          "email-otp-login-controller"
        ],
        "operationId": "verifyEmailOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/email-otp/send": {
      "post": {
        "tags": [
          "email-otp-login-controller"
        ],
        "operationId": "sendEmailOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/verify-otp": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Registration OTP",
        "description": "Verify OTP code to complete user registration. After successful verification, the user account will be activated.",
        "operationId": "verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User verified and activated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "verified"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "error": "invalid_otp"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register User",
        "description": "Register a new user with email verification via OTP. Requires reCAPTCHA v3 token in X-Recaptcha-Token header. An OTP will be sent to the provided email for verification.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "status": "otp_sent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "reCAPTCHA verification failed or invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "error": "recaptcha_failed"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VerifyRequest": {
        "type": "object",
        "description": "OTP verification request with email and OTP code",
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Registration request with name and email",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter JWT token obtained from MagicLink authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}